fix cli arg -c for custom config file
Custom config file not being read if passed-in via "-c" when starting.
For example: bin/netgrasp -c /my/custom/file.conf start

- Fixed -c arg with start
- Minor formatting and cleanup of unused vars, imports, pep8, etc.

Sorry for all of the extra cleanup in netgrasp.py. If it's more than
you care to look over, feel free to just skip it and pull the -c
fix out. It's a small enough change anyway on its own.

I was in there and started cleaning up a few things and then couldn't
stop. :) I use PyCharm and it lints on redundant parens, pep8, etc. So
visually it's distracting, but I think it's still a good idea to cleanup
to aid readability in general. No big deal if you want to ignore this.