general bug fixes to utils/cli.py
So this started out as an attempt to be able to run ng in
the foreground and still send "signals"/commands to it (pidfile).
It ended up not being worth it with subprocesses, pipes, etc.
I.e. just start it, let it daemonize/fork and tail -f.
That's why there's the pep8 "collateral damage" in pretty and
netgrasp.

- Fix invalid var references.
- Refactor module usage/redundant imports.
- Minor formatting.